+++ backend/src/main/java/ru/mephi/trainer/rest/dto/response/TrainerProgressResponse.java
@@ -15,8 +15,8 @@ import java.util.UUID;
 public class TrainerProgressResponse {
     private UUID trainerId;
     private String trainerName;
-    private Double earnedScore;
-    private Double maxPossibleScore;
+    private Integer earnedScore;
+    private Integer maxPossibleScore;
     private Integer tasksCompleted;
     private Integer totalTasks;
     private List<CompletedTaskTrainerPointResponse> tasksInTrainer;
